Keeping 'keyboard' for user debugging
Instead of just catching the ME (m exception error), we add a 'keyboard' command so that the user can debug.

Also, openNSx expects a character vector fot the filename input, AND it expects us to specify that we want to read the signal in microvolts before we say we want to read it at all, e.g. nsxData = utils.openNSx('uv', 'read', char(rawFilePath));